# Conflicts:
#	src/main/java/Business/Actor.java
#	src/main/java/Business/Address.java
#	src/main/java/Business/City.java
#	src/main/java/DAO/ActorDAO.java
#	src/main/java/DAO/AddressDAO.java
#	src/main/java/DAO/CategoryDAO.java
#	src/main/java/DAO/CityDAO.java
#	src/main/java/DAO/FilmDAO.java
#	target/classes/Business/Address.class
#	target/classes/Business/Customer.class
#	target/classes/DAO/AddressDAO.class
#	target/classes/DAO/CustomerDAO.class
#	target/classes/Database/Data.class
#	target/classes/META-INF/persistence.xml
#	target/classes/com/example/sakilagui/HelloApplication.class

+++ src/main/java/DAO/PaymentDAO.java
@@ -16,7 +16,7 @@ public class PaymentDAO implements DAO {
     @Override
     public Optional read(long id) throws IOException, ClassNotFoundException {
         Data data = new Data();
-        Optional<Payment> payment = Optional.ofNullable((Payment) data.getData(Payment.class, (int) id));
+        Optional<Payment> payment = Optional.ofNullable((Payment) data.getData(Payment.class, (short) id));
         return payment;
     }
